pager: address review feedback for less --header support
- Remove static globals; store header_lines/header_width in
  struct child_process instead
- Invert pager_open/pager_open_header: pager_open() now calls
  pager_open_header(0, 0) as a thin wrapper
- Append --header to existing LESS value when set, instead of
  silently skipping header setup
- Flatten nested if-blocks into single setenv() call
- Use int consistently instead of mixing int and size_t
- Document that header_lines=0 means no header regardless of width

+++ include/pager.h
@@ -7,10 +7,8 @@
 #ifndef UTIL_LINUX_PAGER
 #define UTIL_LINUX_PAGER
 
-#include <stddef.h>
-
 void pager_open(void);
-void pager_open_header(int header_lines, size_t first_col_width);
+void pager_open_header(int header_lines, int first_col_width);
 void pager_close(void);
 
 #endif
